# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ecommerce API in go using Gin frame work",
        "title": "E Commerce API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/offer": {
            "get": {
                "description": "Retrieve a list of all available offers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Offer"
                ],
                "summary": "Get a list of offers",
                "operationId": "getOfferList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new offer to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Offer"
                ],
                "summary": "Add a new offer",
                "operationId": "addOffer",
                "parameters": [
                    {
                        "description": "Offer details",
                        "name": "offer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Offer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New Offer Created",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Failed to create offer",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/admin/offer/{ID}": {
            "delete": {
                "description": "Delete an offer from the system by its unique identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Offer"
                ],
                "summary": "Delete an offer by ID",
                "operationId": "deleteOffer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Offer ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted Successfully",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Failed to delete offer",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/admin/ordercancel": {
            "patch": {
                "description": "Allows the admin to cancel an existing order.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Orders"
                ],
                "summary": "Cancel an order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The ID of the order that you want to cancel",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The order has been successfully canceled",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "An error occurred while cancel the order.",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/admin/orders": {
            "get": {
                "description": "Retrieves a list of order items for admin view.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Orders"
                ],
                "summary": "View admin orders",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "404": {
                        "description": "An error occurred while processing your request.",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/admin/orderstatus": {
            "patch": {
                "description": "Allows the admin to update status of an existing order.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Orders"
                ],
                "summary": "Status update of an order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The ID of the order that you want to upfate status",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": " New status for the order",
                        "name": "status",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The order status has been changed successfully ",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "An error occurred while updating status of the order.",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/admin/sales/report/excel": {
            "get": {
                "description": "Generates a sales report in Excel format and sends it as a downloadable file.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Sales"
                ],
                "summary": "Generate sales report in Excel",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": " Internal Server Error",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/admin/sales/report/pdf": {
            "get": {
                "description": "Generates a sales report in Excel format and sends it as a downloadable file.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Sales"
                ],
                "summary": "Generate sales report in Excel",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Offer": {
            "type": "object",
            "properties": {
                "discount": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "offer": {
                    "type": "string"
                },
                "productid": {
                    "type": "integer"
                },
                "valid_from": {
                    "type": "string"
                },
                "valid_to": {
                    "type": "string"
                }
            }
        }
    }
}